:recycle: Restructure Unit Tests
