+++ package.json
@@ -60,6 +60,6 @@
     "build": "gulp build",
     "lint": "gulp lint",
     "prepare": "npm run build",
-    "test": "./node_modules/.bin/mocha --require ts-node/register ./test/**/*.spec.ts"
+    "test": "./node_modules/.bin/mocha --require ts-node/register ./test/*.spec.ts"
   }
 }
